패킷 제너레이터 수정

+++ DH1_Engine/EchoClient/PacketServiceTypeHandler.cpp
@@ -1,7 +1,7 @@
 ﻿#include "pch.h"
 #include "PacketHandler/PacketServiceTypeHandler.h"
 
-bool PacketServiceTypeHandler::HANDLE_SERVICE_TYPE_INVALID(const uint16 size, const uint32 packetId, byte* pBuffer,
+bool PacketServiceTypeHandler::HANDLE_SERVICE_TYPE_INVALID(const uint16 size, const uint16 packetId, byte* pBuffer,
 	PacketSessionRef& pSession)
 {
 	NET_ENGINE_LOG_ERROR("PacketServiceTypeHandler::HANDLE_SERVICE_TYPE_INVALID\n");
